Fix option id autogeneration when results have no id
When we load results from the server if they don't have a dom id we
autogenerate one for them, so we can keep the aria-activedescendant
attribute in the autocomplete input.

When the results element had not an id the autocomplete was assigning
-option-1, -option-2, ids to the results that didn't have ids either.

This changes the behaviour to use as a prefix the results id, if
present, and fallback to 'stimulus-autocomplete' if not.

We also use a static var to ensure the ids are unique in case there
are several autocomplete widgets in the same page.