Respect min-length-value option
When the min length value is set to 0, the autocomplete will not fetch
the data from the server. This is a problem when a consumer wants to
also show something when the input is empty.

For example, it would unblock the following use case:

```html
<input
  name="search"
  data-autocomplete-target="input"
  data-autocomplete-url-value="/search?q={q}"
  data-autocomplete-min-length-value="0"
  data-action="focusin->autocomplete#onInputChange"
/>
```

in which some results are shown as soon as the input is focused.